Bind a shard with setInt rather than narrowing it to a short
A shard is an int everywhere above ShardOwnedStorage - shardCount, the loop
variables, MessageId.shard - while the columns it is bound to are smallint. The
bridge between the two was a (short) cast at every binding: thirty-six of them in
main sources, fourteen more in tests. An external pipeline scanner flags each one
as an unchecked narrowing, and it is right to, in the sense that the casts were
safe only because of an invariant it cannot see - MessageId's own constructor
refuses a shard above Short.MAX_VALUE.

Binding the int directly is not a way of quieting the warning. It is better on
both of the axes that made the cast look necessary.

The index is still used. PostgreSQL's integer_ops btree opfamily carries the
cross-type int2 = int4 operators, so a smallint column compared against an int4
parameter resolves inside the same opfamily. Verified on 17 against a 200k-row
table shaped like the unordered lane: an int2-parameter prepared statement and an
int4 one produce the identical plan, Index Only Scan using the primary key, with
the shard predicate in the Index Cond either way. This mattered enough to check
rather than assume, because readFromCursor runs roughly twice per delivered
message.

The write side gets strictly safer. An out-of-range value now raises "smallint
out of range" from the server instead of wrapping silently, which is what
(short) 32768 did.

queueId is a genuine short and keeps setShort; nothing is cast there. The one
remaining literal, the queue id a pump's storage handle is bound to, becomes
ShardRuntime.NO_QUEUE_ID - a constant expression narrows implicitly on
initialization, where a cast at the call site does not.

In the tests the same change is a loop variable typed short rather than int,
which is what removes the cast at the call instead of relocating it. The
sequence-density test's delivered map moves from Integer to Short keys so its
loops can follow; a short key against an Integer-keyed map would have boxed to
the wrong type and looked up nothing.

Green on 32 unit tests and 59 integration tests covering both lanes, the by-id
admin surface, pull sessions, the registry, transactional enqueue and delayed
delivery, plus ShardOwnedMultiQueueCostIT for the edited map.

+++ components/postgresql-queue-shard-owned/src/main/java/dk/trustworks/essentials/components/queue/shardowned/ShardOwnedSchema.java
@@ -633,7 +633,7 @@ public final class ShardOwnedSchema {
         for (var shard = 0; shard < units; shard++) {
             statement.setShort(1, queueId);
             statement.setString(2, lane);
-            statement.setShort(3, (short) shard);
+            statement.setInt(3, shard);
             statement.addBatch();
         }
     }
